refactor: rename to filter-menu
