+++ src/react/loom-app/option-bar/filter-menu/filter-text-input/index.tsx
